+++ src/app/components/plattform-admin/plattform-admin.component.ts
@@ -3,6 +3,8 @@ import {DataService} from "../../services/DataService";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {StorageService} from "../../services/StorageService";
 import {Router} from "@angular/router";
+import {DeletionConfirmationComponent} from "../deletion-confirmation/deletion-confirmation.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-plattform-admin',
@@ -24,7 +26,7 @@ export class PlattformAdminComponent {
   orgaDescription: string = '';
   toDeleteOrga: string = '';
 
-  constructor(private dataService: DataService, private snackBar: MatSnackBar, private storageService: StorageService, private router: Router) {
+  constructor(private dataService: DataService, private snackBar: MatSnackBar, private storageService: StorageService, private router: Router, private dialog : MatDialog) {
     if (!this.storageService.isPlattformAdmin()) {
       this.router.navigateByUrl('/dashboard');
     }
@@ -63,7 +65,12 @@ export class PlattformAdminComponent {
     if (orgId == '') {
       this.snackBar.open('Bitte geben Sie eine Organisations-ID an', 'OK', {duration: 3000});
     } else {
-      this.dataService.deleteOrganization(orgId).subscribe(() => this.snackBar.open('Organisation gelöscht', 'OK', {duration: 3000}));
+      const dialogRef = this.dialog.open(DeletionConfirmationComponent,{});
+      dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+        if (confirmed) {
+          this.dataService.deleteOrganization(orgId).subscribe(() => this.snackBar.open('Organisation gelöscht', 'OK', {duration: 3000}));
+        }
+      });
       this.toDeleteOrga = '';
     }
   }
@@ -72,13 +79,23 @@ export class PlattformAdminComponent {
    * Calls the DataService to delete all entries of the database
    */
   clearDB() {
-    this.dataService.deleteDatabase().subscribe(() => this.snackBar.open('Datenbank geleert', 'OK', {duration: 3000}));
+    const dialogRef = this.dialog.open(DeletionConfirmationComponent,{});
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.dataService.deleteDatabase().subscribe(() => this.snackBar.open('Datenbank geleert', 'OK', {duration: 3000}));
+      }
+    });
   }
 
   /**
    * Calls the DataService to re-initialize the database
    */
   resetDB() {
-    this.dataService.resetDatabase().subscribe(() => this.snackBar.open('Datenbank zurückgesetzt', 'OK', {duration: 3000}));
+    const dialogRef = this.dialog.open(DeletionConfirmationComponent,{});
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.dataService.resetDatabase().subscribe(() => this.snackBar.open('Datenbank zurückgesetzt', 'OK', {duration: 3000}));
+      }
+    });
   }
 }
